refactor(pytest): putting the setup in conftest before imports simplifies
