--- server/pytest.ini
@@ -1,3 +1,0 @@
-[pytest]
-# this ensures FLASK_ENV = 'testing' is on even before the top imports of conftest
-python_files = test_environment.py test_*.py
